+++ NEWS.md
@@ -7,6 +7,7 @@
 * New function `catalog_retile` supersede the function `catalog_reshape` being able to do the same and much more.
 * New function `lastrees_ptree` with Vega et al. (2014) algorithm.
 * New function `lastrees_hamraz` with Hamraz et al. (2016) algorithm.
+* New function `tree_detection_manual`for manual correction of tree detection.
 
 #### ENHANCEMENTS